Trekk ut interface til egen fil
