+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vedtak/begrunnelser/Standardbegrunnelse.kt
@@ -1,7 +1,5 @@
 package no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser
 
-import com.fasterxml.jackson.annotation.JsonSubTypes
-import com.fasterxml.jackson.annotation.JsonTypeInfo
 import no.nav.familie.ba.sak.common.Feil
 import no.nav.familie.ba.sak.common.NullablePeriode
 import no.nav.familie.ba.sak.common.Utils.konverterEnumsTilString
@@ -14,34 +12,6 @@ import no.nav.familie.ba.sak.kjerne.vedtak.domene.MinimertRestPerson
 import no.nav.familie.ba.sak.kjerne.vedtak.domene.hentRelevanteEndringsperioderForBegrunnelse
 import javax.persistence.AttributeConverter
 import javax.persistence.Converter
-
-@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "vedtakBegrunnelseType")
-@JsonSubTypes(
-    JsonSubTypes.Type(value = Standardbegrunnelse::class, name = "INNVILGET"),
-    JsonSubTypes.Type(value = Standardbegrunnelse::class, name = "REDUKSJON"),
-    JsonSubTypes.Type(value = Standardbegrunnelse::class, name = "AVSLAG"),
-    JsonSubTypes.Type(value = Standardbegrunnelse::class, name = "OPPHØR"),
-    JsonSubTypes.Type(value = Standardbegrunnelse::class, name = "FORTSATT_INNVILGET"),
-    JsonSubTypes.Type(value = Standardbegrunnelse::class, name = "ENDRET_UTBETALING"),
-    JsonSubTypes.Type(value = Standardbegrunnelse::class, name = "ETTER_ENDRET_UTBETALING"),
-    JsonSubTypes.Type(value = EØSStandardbegrunnelse::class, name = "EØS_INNVILGET"),
-    JsonSubTypes.Type(value = EØSStandardbegrunnelse::class, name = "EØS_AVSLAG"),
-    JsonSubTypes.Type(value = EØSStandardbegrunnelse::class, name = "EØS_OPPHØR"),
-)
-interface IVedtakBegrunnelse {
-
-    val sanityApiNavn: String
-    val vedtakBegrunnelseType: VedtakBegrunnelseType
-    val kanDelesOpp: Boolean
-
-    fun delOpp(
-        restBehandlingsgrunnlagForBrev: RestBehandlingsgrunnlagForBrev,
-        triggesAv: TriggesAv,
-        periode: NullablePeriode
-    ): List<BrevBegrunnelseGrunnlagMedPersoner>
-
-    fun enumnavnTilString(): String
-}
 
 val hjemlerTilhørendeFritekst = setOf(2, 4, 11)
